requirements.txt
onnxruntime-genai-directml
flask